attendance_project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-barcode[images]>=0.15.0
reportlab>=4.0.0
qrcode[pil]>=7.4.0